Update fixtures
Added a fabfile command, removed lol.yaml, and renamed initial_data's
extension {yaml,json} because it's not really YAML (although it seemed
to work before). Also added some new pages and stuff.

Things to consider: using custom templates or just saving the content
on the filesystem and maybe caching it in the DB?