Update payment page, by request
Made the top navigation bar horizontal rather than vertical. I
initially wanted to put in the '::' using CSS (:after) but it proved
to be too much trouble because of the initial li and anyway this method
probably has better browser support.

Also added a rudimentary js "fee calculator". Was kind of fun. This
required adding a template filter for ranges.